+++ CHANGELOG.md
@@ -10,6 +10,11 @@ List all changes after the last release here (newer on top). Each change on a se
 
 ### Added
 
+- Admin: Ability to login as the selected contact if it's a user
+
+### Fixed
+
+- Admin: Now when activating/deactivating user it's contact will also change
 - Admin: New notification for when a account get's reactivated
 
 ## [1.10.11] - 2020-04-23